Added privacy test
added privacy test and a test for preprocessing
