+++ requirements.txt
@@ -1,6 +1,6 @@
 tensorflow==2.16.1
 boto3==1.34.49
-keras==3.0.0
+keras~=3.3.3
 numpy==1.24.3
 scikit-learn==1.3.2
 python-dotenv==1.0.1
@@ -12,4 +12,8 @@ lib-ml-remla-team14-a2==1.0.14.dev0
 typing-extensions==4.5.0
 scmrepo==0.0.25
 pydantic==1.9.0
-pytest
+pytest~=8.2.1
+botocore~=1.34.69
+pandas~=2.1.4
+seaborn~=0.13.2
+matplotlib~=3.8.3
